pyproject.toml
[project]
name = "mal_cord"
description = "Goldy Bot extension for browsing ⛩️ anime."
authors = [
    {name = "Goldy", email = "goldy@devgoldy.xyz"},
]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = [
    "anime", "myanimelist", "mal"
]
dependencies = [
    "GoldyBot>=5.0dev10",

    "jikanpy-v4",
    "Pillow"
]
version = "1.1"

[project.optional-dependencies]
dev = [
    "ruff"
]

[project.urls]
GitHub = "https://github.com/THEGOLDENPRO/mal_cord"
BugTracker = "https://github.com/THEGOLDENPRO/mal_cord/issues"



# Ruff
# ---------------

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = ["E501", "E402", "E702", "E701", "E731"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "env",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py38"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.per-file-ignores]
"__init__.py" = []